bot/requirements.txt
python-telegram-bot==20.0
python-dotenv==0.21.0
Flask==2.2.2
nest_asyncio==1.6.0